[UX] Show Message When Movie List Retrieval Fails
The "Failed to Retrieve, Swipe to Retry" message will be shown instead of the RecyclerView when movie list is empty. Movie list is empty when there was no prior successful retrieval so the cache is empty. Otherwise, if there movies cached in database, a "Failed to Retrieve, Swipe to Retry" message will be shown using a Snackbar whilst displaying the cached movies.